requirements.txt
# Core dependencies for production deployment
# This file serves as a backup for environments where uv might not work properly

# Web framework
fastapi==0.115.14
uvicorn[standard]==0.34.3

# MCP and Agent frameworks
fast-agent-mcp[cli]==0.2.35
fastmcp==2.9.2

# Database and storage
supabase==2.16.0

# HTTP client
httpx==0.28.1

# Template engine
jinja2==3.1.6

# Async file operations
aiofiles==24.1.0

# Environment management
python-dotenv>=1.0.0

# Type hints
typing-extensions>=4.8.0

# Development tools (optional for production)
# pytest>=8.4.1
# pytest-asyncio>=1.0.0
# black>=23.0.0
# ruff>=0.1.0